lesson11 homework

+++ src/main/java/com/tbezdetnaya/app/homework/lesson9/ArrayGenerator.java
@@ -17,14 +17,12 @@ public class ArrayGenerator {
         }
         return arr;
     }
-    public Integer[] createArray (){
-        Integer arr[] = new Integer[n];
+    public double[] createArray (){
+        double arr[] = new double[n];
         for (int i = 0; i < arr.length; i++) {
             arr[i] = (int) (Math.random() * 100);
         }
         return arr;
     }
-
-
 
 }
